# Switch regex for delete_suffix in normalize_path